chore(engine): extract stream events to dedicated file
