+++ engine/internal/types/testdata/contracts.json
@@ -149,6 +149,7 @@
     "message",
     "messagesAfter",
     "messagesBefore",
+    "metadata",
     "method",
     "options",
     "partialInput",
